requirements.txt
faker
kafka-python
pyspark
psycopg2-binary